+++ src/sidebar/index.js
@@ -184,30 +184,30 @@ module.exports = angular.module('h', [
   .directive('spinner', require('./directive/spinner'))
   .directive('windowScroll', require('./directive/window-scroll'))
 
-  .service('analytics', require('./analytics'))
-  .service('annotationMapper', require('./annotation-mapper'))
-  .service('annotationUI', require('./annotation-ui'))
+  .service('analytics', require('./services/analytics'))
+  .service('annotationMapper', require('./services/annotation-mapper'))
+  .service('annotationUI', require('./services/annotation-ui'))
   .service('api', require('./services/api'))
-  .service('apiRoutes', require('./api-routes'))
-  .service('auth', require('./oauth-auth'))
+  .service('apiRoutes', require('./services/api-routes'))
+  .service('auth', require('./services/oauth-auth'))
   .service('bridge', require('../shared/bridge'))
-  .service('drafts', require('./drafts'))
-  .service('features', require('./features'))
-  .service('flash', require('./flash'))
-  .service('frameSync', require('./frame-sync').default)
-  .service('groups', require('./groups'))
-  .service('localStorage', require('./local-storage'))
-  .service('permissions', require('./permissions'))
-  .service('queryParser', require('./query-parser'))
-  .service('rootThread', require('./root-thread'))
-  .service('searchFilter', require('./search-filter'))
-  .service('serviceUrl', require('./service-url'))
-  .service('session', require('./session'))
-  .service('streamer', require('./streamer'))
-  .service('streamFilter', require('./stream-filter'))
-  .service('tags', require('./tags'))
-  .service('unicode', require('./unicode'))
-  .service('viewFilter', require('./view-filter'))
+  .service('drafts', require('./services/drafts'))
+  .service('features', require('./services/features'))
+  .service('flash', require('./services/flash'))
+  .service('frameSync', require('./services/frame-sync').default)
+  .service('groups', require('./services/groups'))
+  .service('localStorage', require('./services/local-storage'))
+  .service('permissions', require('./services/permissions'))
+  .service('queryParser', require('./services/query-parser'))
+  .service('rootThread', require('./services/root-thread'))
+  .service('searchFilter', require('./services/search-filter'))
+  .service('serviceUrl', require('./services/service-url'))
+  .service('session', require('./services/session'))
+  .service('streamer', require('./services/streamer'))
+  .service('streamFilter', require('./services/stream-filter'))
+  .service('tags', require('./services/tags'))
+  .service('unicode', require('./services/unicode'))
+  .service('viewFilter', require('./services/view-filter'))
 
   .value('Discovery', require('../shared/discovery'))
   .value('ExcerptOverflowMonitor', require('./util/excerpt-overflow-monitor'))
