+++ src/sidebar/services/analytics.js
