+++ src/sidebar/services/features.js
@@ -11,8 +11,8 @@
  */
 'use strict';
 
-var events = require('./events');
-var bridgeEvents = require('../shared/bridge-events');
+var events = require('../events');
+var bridgeEvents = require('../../shared/bridge-events');
 
 
 // @ngInject
