+++ src/sidebar/services/flash.js
