+++ src/sidebar/services/local-storage.js
