+++ src/sidebar/services/oauth-auth.js
@@ -1,8 +1,8 @@
 'use strict';
 
-var events = require('./events');
-var resolve = require('./util/url-util').resolve;
-var serviceConfig = require('./service-config');
+var events = require('../events');
+var resolve = require('../util/url-util').resolve;
+var serviceConfig = require('../service-config');
 
 /**
  * @typedef RefreshOptions
