+++ src/sidebar/services/query-parser.coffee
