+++ src/sidebar/services/search-filter.js
