+++ src/sidebar/services/stream-filter.coffee
