+++ src/sidebar/services/test/groups-test.js
@@ -1,9 +1,9 @@
 'use strict';
 
-var events = require('../events');
-var fakeReduxStore = require('./fake-redux-store');
+var events = require('../../events');
+var fakeReduxStore = require('../../test/fake-redux-store');
 var groups = require('../groups');
-var unroll = require('../../shared/test/util').unroll;
+var unroll = require('../../../shared/test/util').unroll;
 
 // Return a mock session service containing three groups.
 var sessionWithThreeGroups = function() {
