+++ src/sidebar/services/test/local-storage-test.js
