+++ src/sidebar/services/test/permissions-test.js
