+++ src/sidebar/services/test/session-test.js
@@ -2,7 +2,7 @@
 
 var angular = require('angular');
 
-var events = require('../events');
+var events = require('../../events');
 
 var mock = angular.mock;
