+++ src/sidebar/services/test/unicode-test.js
