+++ src/sidebar/services/test/view-filter-test.coffee
