+++ src/sidebar/services/unicode.js
