+++ src/sidebar/services/view-filter.js
